Extract SId helpers
